pyproject.toml
[project]
name = "bot-nfse-nacional"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "patchright>=1.56.0",
    "pdfplumber>=0.11.8",
    "playwright>=1.56.0",
    "pyautogui>=0.9.54",
    "python-telegram-bot>=22.5",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
]
